Refactor the Process Plugin into a service that we can use in other classes. Update the accordion importer to now pass the body value over to the Media Embed Service to get thew new media embed code.
